Restructure "kinetics orgs invite" to "kinetics orgs members invite"
