+++ cli/src/commands/orgs.rs
@@ -1,11 +1,12 @@
 pub mod create;
 pub mod delete;
-pub mod invite;
 pub mod list;
+pub mod members;
+
 use crate::commands::orgs::create::CreateCommand;
 use crate::commands::orgs::delete::DeleteCommand;
-use crate::commands::orgs::invite::InviteCommand;
 use crate::commands::orgs::list::ListCommand;
+use crate::commands::orgs::members::MembersCommands;
 use clap::Subcommand;
 
 #[derive(Subcommand)]
@@ -16,9 +17,10 @@ pub(crate) enum OrgsCommands {
     /// Delete an org
     Delete(DeleteCommand),
 
-    /// Invite a member to an org (sends an email)
-    Invite(InviteCommand),
-
     /// List all orgs you belong to
     List(ListCommand),
+
+    /// Manage org members
+    #[clap(subcommand)]
+    Members(MembersCommands),
 }
